Don't try to destroy a moved from buffer. Also set the correct descriptor write type.

+++ Source/Core/Buffer.cpp
@@ -32,7 +32,7 @@ Buffer::Buffer(RenderDevice* dev,
 Buffer::~Buffer()
 {
     const bool shouldDestroy = release();
-    if(shouldDestroy)
+    if(shouldDestroy && mBuffer != vk::Buffer{nullptr})
         getDevice()->destroyBuffer(*this);
 }
